requirements.txt
alembic==1.7.5
Babel==2.9.1
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
decorator==5.1.1
docxcompose==1.3.4
docxtpl==0.15.2
dominate==2.6.0
Flask==2.0.2
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
greenlet==1.1.2
gspread==5.1.1
gspread-pandas==3.0.4
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.1
oauthlib==3.1.1
pandas==1.3.5
passlib==1.7.4
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.30
uritemplate==4.1.1
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
